[backend] refactor isStoreRelationPir
